config simplify
