--- web/apps/wps-web/src/features/landingPage/landingPageConfig.ts
@@ -1,13 +1,0 @@
-export const BCPS_SECTION_ID = 'bcps-access-only-heading'
-export const FAVOURITES_SECTION_ID = 'favourites-heading'
-export const PUBLIC_SECTION_ID = 'public-access-heading'
-export const LANDING_PAGE_FAVOURITES_STORAGE_KEY = 'wps-landing-page-favourites'
-
-export const ICON_TILE_RADIUS = '14px'
-export const SECTION_RADIUS = '18px'
-export const TOOL_ROW_RADIUS = '16px'
-export const FAVOURITES_COLOUR = '#3f743f'
-export const PUBLIC_TOOL_ICON_COLOUR = '#FF6900'
-
-export const BCWS_PREDICTIVE_SERVICES_EMAIL = 'BCWS.PredictiveServices@gov.bc.ca'
-export const TECH_SERVICES_EMAIL = 'BCWS.TechServices@gov.bc.ca'
