+++ web/packages/ui/src/theme.ts
@@ -84,8 +84,7 @@ export const fireTableTheme = createTheme({
 })
 
 export const BACKGROUND_COLOR = { backgroundColor: '#e9ecf5' }
-
-export const LANDING_BACKGROUND_COLOUR = '#f2f2f2'
+export const PUBLIC_TOOL_ICON_COLOUR = '#FF6900'
 
 export const PLANNING_AREA = {
   minWidth: 45,
